Move variable operations to file.
