+++ core/CMakeLists.txt
@@ -34,6 +34,7 @@ set(SRC_FILES
     variable_binary_arithmetic.cpp
     variable_instantiate_basic.cpp
     variable_instantiate_dataset.cpp
+    variable_operations.cpp
     view_index.cpp)
 
 set(LINK_TYPE "STATIC")
